+++ test/rag/retriever/test_faiss_retriever.py
@@ -1,7 +1,7 @@
 import pytest 
 
-from vectoria_lib.rag.retriever.faiss_retriever import FaissRetriever
-from vectoria_lib.rag.vector_store.faiss_vector_store import FaissVectorStore
+from vectoria_lib.components.retriever.faiss_retriever import FaissRetriever
+from vectoria_lib.components.vector_store.faiss_vector_store import FaissVectorStore
 
 @pytest.mark.parametrize("k",[1,2,3])
 def test_faiss_retriever(k, config, index_test_folder):
