+++ vectoria_lib/applications/application_builder.py
@@ -7,14 +7,14 @@ import logging
 
 from vectoria_lib.applications.qa import QAApplication
 from vectoria_lib.common.config import Config
-from vectoria_lib.rag.postretrieval_steps.full_paragraphs import FullParagraphs
-from vectoria_lib.rag.vector_store.vectore_store_builder import VectorStoreBuilder
-from vectoria_lib.rag.retriever.retriever_builder import RetrieverBuilder
+from vectoria_lib.components.postretrieval_steps.full_paragraphs import FullParagraphs
+from vectoria_lib.components.vector_store.vectore_store_builder import VectorStoreBuilder
+from vectoria_lib.components.retriever.retriever_builder import RetrieverBuilder
 from vectoria_lib.llm.llm_factory import LLMFactory
 from vectoria_lib.applications.chains import create_qa_chain
 from vectoria_lib.applications.prompt_builder import PromptBuilder
 from vectoria_lib.llm.parser import CustomResponseParser
-from vectoria_lib.rag.postretrieval_steps.huggingface_reranker import Reranker
+from vectoria_lib.components.postretrieval_steps.huggingface_reranker import Reranker
 
 class ApplicationBuilder:
